refactor: 패키지 변경 및 DTO생성
