+++ backend/src/main/java/org/donggle/backend/ui/response/AccessTokenResponse.java
@@ -1,0 +1,6 @@
+package org.donggle.backend.ui.response;
+
+public record AccessTokenResponse(
+        String accessToken
+) {
+}
