+++ backend/src/main/java/org/donggle/backend/ui/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package org.donggle.backend.auth;
+package org.donggle.backend.ui.response;
 
 public record TokenResponse(String accessToken, String refreshToken) {
 }
